requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
gunicorn==22.0.0
python-telegram-bot>=20.0,<22.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
python-escpos>=3.0
